ZA-021: Added SEO to reset password page
